Fixed pixtoCoords to work with both boards based on turn
pygameloop: changed pixtocoords function to track turn so the right board is targeted based on turn. Probably could add some kind of constant display of coords targeted to help debugging and gameplay
board: ran into error with isSunk being called with too many arguments, isSunk on ship class is just a method on itself, nothing needs to be passed in.